requirements.txt
absl-py==2.3.1
amqp==5.3.1
annotated-types==0.7.0
arabic-reshaper==3.0.0
asgiref==3.9.2
asn1crypto==1.5.1
astunparse==1.6.3
async-timeout==5.0.1
billiard==4.2.2
blis==1.3.0
cachetools==5.5.2
catalogue==2.0.10
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpathlib==0.23.0
colorama==0.4.6
confection==0.1.5
contourpy==1.3.2
cryptography==46.0.2
cssselect2==0.8.0
cycler==0.12.1
cymem==2.0.11
Django==5.2.6
djangorestframework==3.16.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
flatbuffers==25.9.23
fonttools==4.60.0
freetype-py==2.5.1
gast==0.4.0
google-auth==2.40.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.75.1
h5py==3.14.0
html5lib==1.1
idna==3.10
imbalanced-learn==0.14.0
Jinja2==3.1.6
joblib==1.5.2
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.9
kombu==5.5.4
langcodes==3.5.0
language_data==1.3.0
libclang==18.1.1
lxml==6.0.2
marisa-trie==1.3.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mdurl==0.1.2
murmurhash==1.0.13
nltk==3.9.2
numpy==1.23.5
oauthlib==3.3.1
opt_einsum==3.4.0
oscrypto==1.3.0
packaging==25.0
pandas==2.3.2
pillow==11.3.0
preshed==3.0.10
prompt_toolkit==3.0.52
protobuf==3.19.6
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.28.0
pycparser==2.23
pydantic==2.12.2
pydantic_core==2.41.4
Pygments==2.19.2
pyHanko==0.31.0
pyhanko-certvalidator==0.29.0
pyparsing==3.2.5
pypdf==6.1.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
redis==6.4.0
regex==2025.9.18
reportlab==4.4.4
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rlPyCairo==0.4.0
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.15.3
shellingham==1.5.4
six==1.17.0
smart_open==7.3.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.3
srsly==2.5.1
svglib==1.6.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-cpu==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.10.0
termcolor==3.1.0
thinc==8.3.6
threadpoolctl==3.6.0
tinycss2==1.4.0
tqdm==4.67.1
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
vine==5.1.0
wasabi==1.1.3
wcwidth==0.2.14
weasel==0.4.1
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.17.3
xgboost==3.0.5
xhtml2pdf==0.2.17
